Added Script for player lead decoding
I got most of the infrastructure for the dictionary system I'm planning to use for integration with the Translator.cs script, and most of what I wrote does work independently of the whole (that I know of), however there was one error on line 53 that prevented the cipher text sub menus of each English letter from opening properly, a "NullReferenceException." I troubleshooted it for about 2 hours at the end of my 7 hour programming session, but to no avail, so I'll have to revisit this to get it working. That was my last issue to fix before moving on to work on the integration with the Translator.cs script though that I know of, so hopefully it shouldn't take too much longer.